requirements.txt
pyTelegramBotAPI
SQLAlchemy
psycopg2-binary
python-dotenv